Worked on the Back End Framework
Made it so the back end would take information from the browser and store it on the database. Also tested it with postman as the mock front end framework and send  information as a key value pair and sent the post request and it correctly created the user in the database.